consolidate the tiny differences between work, image and collection factories making object factory do the right thing in all cases, and removing the selector
