+++ app/models/concerns/bulkrax/import_behavior.rb
@@ -54,7 +54,7 @@ module Bulkrax
     end
 
     def factory
-      @factory ||= Bulkrax::ApplicationFactory.for(factory_class.to_s).new(self.parsed_metadata, identifier, parser.files_path, [], user)
+      @factory ||= Bulkrax::ObjectFactory.new(self.parsed_metadata, identifier, parser.files_path, [], user, factory_class)
     end
 
     def factory_class
